Maintain a installation counter.
- Extact node_module folder only for the first installation.
- Remove node_module folder after the last installation removed.

+++ Nancy.ViewEngines.React/tools/uninstall.ps1
@@ -1,4 +1,31 @@
-﻿# TODO it will break when several projects in a solution install this package
 param($installPath, $toolsPath, $package, $project)
-$nodeModulePath = (join-path $toolsPath "../node_modules")
-& cmd /c rmdir /S /Q "$nodeModulePath" # Use CMD is to avoid path too long issue
+
+# Get installation counter to calculate if remove node_modules folder
+$remove = $false
+$countFile = (Join-Path $toolsPath "count.txt")
+if (-not [IO.File]::Exists($countFile))
+{
+  $remove = $true
+}
+else
+{
+  $count = +[IO.File]::ReadAllText($countFile)
+  if ($count -eq 1)
+  {
+    $remove = $true
+    [IO.File]::Delete($countFile)
+  }
+  else
+  {
+    [IO.File]::WriteAllText($countFile, $count - 1)
+  }
+}
+
+if ($remove)
+{
+  # Remove node_modules folder
+  $nodeModulePath = (Join-Path $toolsPath "../node_modules")
+
+  # Use CMD to avoid path too long issue. See http://stackoverflow.com/a/18267957/1436671
+  & cmd /c rmdir /S /Q "$nodeModulePath"
+}
